Remove unnecessary Context struct
